refactor: Order 쿼리 코드 Fulfillment 하위로 이동
