+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/product/ProductRepository.java
@@ -1,9 +1,10 @@
 package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.product;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<ProductRepository, ProductNumber> {
+public interface ProductRepository extends CrudRepository<Product, ProductNumber> {
 
-    Priority getLowestPriory();
+//    Priority getLowestPriory();
 
 }
