+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/query/dto/OrderDto.java
@@ -1,7 +1,6 @@
-package kr.tatine.manibogo_oms_v2.order.query.dto;
+package kr.tatine.manibogo_oms_v2.fulfillment.query.dto;
 
 import kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order.OrderLocation;
-import kr.tatine.manibogo_oms_v2.order.command.domain.OrderState;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
