+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/query/dto/OrderListDto.java
@@ -1,0 +1,16 @@
+package kr.tatine.manibogo_oms_v2.fulfillment.query.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import java.util.List;
+
+@Getter
+@Setter
+@ToString
+public class OrderListDto {
+
+    private List<OrderDto> orders;
+
+}
