+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/query/dto/OrderState.java
@@ -1,4 +1,4 @@
-package kr.tatine.manibogo_oms_v2.order.command.domain;
+package kr.tatine.manibogo_oms_v2.fulfillment.query.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
