+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/ui/EditOrderListForm.java
@@ -1,4 +1,4 @@
-package kr.tatine.manibogo_oms_v2.order.query.dto;
+package kr.tatine.manibogo_oms_v2.fulfillment.ui;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,8 +9,8 @@ import java.util.List;
 @Getter
 @Setter
 @ToString
-public class OrderListDto {
+public class EditOrderListForm {
 
-    private List<OrderDto> orders;
+    private List<EditOrderForm> orders;
 
 }
